+++ documentation/Snake-AI-Doc.blg
@@ -6,46 +6,58 @@ Reallocating 'name_of_file' (item size: 1) to 6 items.
 The style file: plain.bst
 Database file #1: bibliography.bib
 Warning--entry type for "geeksdijkstra" isn't style-file defined
---line 8 of file bibliography.bib
-You've used 1 entry,
+--line 3 of file bibliography.bib
+Warning--entry type for "classLoader" isn't style-file defined
+--line 21 of file bibliography.bib
+Warning--entry type for "Cloneable" isn't style-file defined
+--line 35 of file bibliography.bib
+Warning--entry type for "awt" isn't style-file defined
+--line 49 of file bibliography.bib
+Warning--entry type for "runnable" isn't style-file defined
+--line 61 of file bibliography.bib
+Warning--to sort, need author or key in classLoader
+Warning--to sort, need author or key in runnable
+Warning--to sort, need author or key in Cloneable
+Warning--to sort, need author or key in awt
+You've used 5 entries,
             2118 wiz_defined-function locations,
-            501 strings with 4110 characters,
-and the built_in function-call counts, 217 in all, are:
-= -- 17
-> -- 7
+            512 strings with 4238 characters,
+and the built_in function-call counts, 825 in all, are:
+= -- 69
+> -- 11
 < -- 0
-+ -- 3
++ -- 7
 - -- 2
-* -- 8
-:= -- 41
-add.period$ -- 2
-call.type$ -- 1
-change.case$ -- 4
+* -- 32
+:= -- 101
+add.period$ -- 6
+call.type$ -- 5
+change.case$ -- 16
 chr.to.int$ -- 0
-cite$ -- 1
-duplicate$ -- 7
-empty$ -- 23
+cite$ -- 9
+duplicate$ -- 35
+empty$ -- 115
 format.name$ -- 2
-if$ -- 44
+if$ -- 196
 int.to.chr$ -- 0
-int.to.str$ -- 1
+int.to.str$ -- 5
 missing$ -- 0
-newline$ -- 7
+newline$ -- 19
 num.names$ -- 2
-pop$ -- 9
+pop$ -- 45
 preamble$ -- 1
-purify$ -- 3
+purify$ -- 11
 quote$ -- 0
-skip$ -- 8
+skip$ -- 40
 stack$ -- 0
-substring$ -- 5
-swap$ -- 1
+substring$ -- 25
+swap$ -- 5
 text.length$ -- 0
 text.prefix$ -- 0
 top$ -- 0
-type$ -- 4
-warning$ -- 0
+type$ -- 20
+warning$ -- 4
 while$ -- 2
-width$ -- 2
-write$ -- 10
-(There was 1 warning)
+width$ -- 6
+write$ -- 34
+(There were 9 warnings)
