+++ documentation/Snake-AI-Doc.out
@@ -25,14 +25,18 @@
 \BOOKMARK [1][]{section.33}{\376\377\0008\000\040\000O\000u\000r\000\040\000B\000o\000t\000s\000\040\000w\000i\000t\000h\000\040\000d\000i\000f\000f\000e\000r\000e\000n\000t\000\040\000i\000m\000p\000l\000e\000m\000e\000n\000t\000a\000t\000i\000o\000n\000s}{}% 26
 \BOOKMARK [2][]{subsection.34}{\376\377\0008\000.\0001\000\040\000B\000o\000t\0001}{section.33}% 27
 \BOOKMARK [2][]{subsection.35}{\376\377\0008\000.\0002\000\040\000B\000o\000t\0002}{section.33}% 28
-\BOOKMARK [2][]{subsection.36}{\376\377\0008\000.\0003\000\040\000B\000o\000t\0003}{section.33}% 29
+\BOOKMARK [2][]{subsection.36}{\376\377\0008\000.\0003\000\040\000S\000i\000m\000p\000l\000i\000f\000i\000e\000d\000D\000i\000j\000k\000s\000t\000r\000a\000V\0002}{section.33}% 29
 \BOOKMARK [1][]{section.38}{\376\377\0009\000\040\000E\000x\000p\000e\000r\000i\000m\000e\000n\000t\000a\000l\000\040\000R\000e\000s\000u\000l\000t\000s\000\040\000a\000n\000d\000\040\000S\000t\000a\000t\000i\000s\000t\000i\000c\000a\000l\000\040\000T\000e\000s\000t\000s}{}% 30
-\BOOKMARK [2][]{subsection.39}{\376\377\0009\000.\0001\000\040\000S\000i\000m\000u\000l\000a\000t\000i\000o\000n\000s\000\040\000\050\000P\000l\000a\000y\000\040\000w\000i\000t\000h\000\040\000t\000h\000e\000\040\000P\000a\000r\000a\000m\000e\000t\000e\000r\000s\000!\000\051}{section.38}% 31
-\BOOKMARK [2][]{subsection.40}{\376\377\0009\000.\0002\000\040\000U\000s\000e\000\040\000B\000e\000n\000c\000h\000m\000a\000r\000k\000s}{section.38}% 32
-\BOOKMARK [2][]{subsection.41}{\376\377\0009\000.\0003\000\040\000T\000a\000b\000l\000e\000s\000:\000\040\000i\000n\000p\000u\000t\000\040\000d\000a\000t\000a\000\040\000d\000e\000t\000a\000i\000l\000s\000,\000\040\000r\000e\000s\000u\000l\000t\000s\000\040\000o\000f\000\040\000d\000i\000f\000f\000e\000r\000e\000n\000t\000\040\000a\000l\000g\000o\000r\000i\000t\000h\000m\000s}{section.38}% 33
-\BOOKMARK [2][]{subsection.42}{\376\377\0009\000.\0004\000\040\000C\000h\000a\000r\000t\000s}{section.38}% 34
-\BOOKMARK [2][]{subsection.43}{\376\377\0009\000.\0005\000\040\000E\000v\000a\000l\000u\000a\000t\000i\000o\000n\000s}{section.38}% 35
-\BOOKMARK [1][]{section.44}{\376\377\0001\0000\000\040\000C\000o\000n\000c\000l\000u\000s\000i\000o\000n\000s\000\040\000a\000n\000d\000\040\000F\000u\000t\000u\000r\000e\000\040\000W\000o\000r\000k}{}% 36
-\BOOKMARK [2][]{subsection.45}{\376\377\0001\0000\000.\0001\000\040\000H\000o\000w\000\040\000t\000h\000e\000\040\000t\000e\000a\000m\000\040\000w\000o\000r\000k\000\040\000w\000e\000n\000t}{section.44}% 37
-\BOOKMARK [2][]{subsection.46}{\376\377\0001\0000\000.\0002\000\040\000W\000h\000a\000t\000\040\000w\000e\000\040\000h\000a\000v\000e\000\040\000l\000e\000a\000r\000n\000e\000d}{section.44}% 38
-\BOOKMARK [2][]{subsection.47}{\376\377\0001\0000\000.\0003\000\040\000I\000d\000e\000a\000s\000\040\000f\000o\000r\000\040\000t\000h\000e\000\040\000f\000u\000t\000u\000r\000e\000\040\000d\000e\000v\000e\000l\000o\000p\000m\000e\000n\000t\000\040\000o\000f\000\040\000y\000o\000u\000r\000\040\000a\000p\000p\000l\000i\000c\000a\000t\000i\000o\000n\000,\000\040\000n\000e\000w\000\040\000a\000l\000g\000o\000r\000i\000t\000h\000m\000s}{section.44}% 39
+\BOOKMARK [2][]{subsection.39}{\376\377\0009\000.\0001\000\040\000S\000i\000m\000u\000l\000a\000t\000i\000o\000n\000s}{section.38}% 31
+\BOOKMARK [2][]{subsection.42}{\376\377\0009\000.\0002\000\040\000R\000e\000s\000u\000l\000t\000s\000\040\000o\000f\000\040\000D\000i\000f\000f\000e\000r\000e\000n\000t\000\040\000A\000l\000g\000o\000r\000i\000t\000h\000m\000s}{section.38}% 32
+\BOOKMARK [3][]{subsubsection.43}{\376\377\0009\000.\0002\000.\0001\000\040\000R\000e\000c\000u\000r\000s\000i\000v\000e\000B\000o\000t\000\040\000v\000s\000.\000\040\000S\000i\000m\000p\000l\000i\000f\000i\000e\000d\000D\000i\000j\000k\000s\000t\000r\000a}{subsection.42}% 33
+\BOOKMARK [3][]{subsubsection.45}{\376\377\0009\000.\0002\000.\0002\000\040\000R\000e\000c\000u\000r\000s\000i\000v\000e\000B\000o\000t\000\040\000v\000s\000.\000\040\000S\000i\000m\000p\000l\000i\000f\000i\000e\000d\000D\000i\000j\000k\000s\000t\000r\000a\000V\0002}{subsection.42}% 34
+\BOOKMARK [3][]{subsubsection.47}{\376\377\0009\000.\0002\000.\0003\000\040\000O\000t\000h\000e\000r\000\040\000R\000e\000s\000u\000l\000t\000s}{subsection.42}% 35
+\BOOKMARK [2][]{subsection.50}{\376\377\0009\000.\0003\000\040\000C\000h\000a\000r\000t\000s}{section.38}% 36
+\BOOKMARK [3][]{subsubsection.51}{\376\377\0009\000.\0003\000.\0001\000\040\000R\000e\000c\000u\000r\000s\000i\000v\000e\000B\000o\000t\000\040\000v\000s\000.\000\040\000S\000i\000m\000p\000l\000i\000f\000i\000e\000d\000D\000i\000j\000k\000s\000t\000r\000a}{subsection.50}% 37
+\BOOKMARK [3][]{subsubsection.54}{\376\377\0009\000.\0003\000.\0002\000\040\000R\000e\000c\000u\000r\000s\000i\000v\000e\000B\000o\000t\000\040\000v\000s\000.\000\040\000S\000i\000m\000p\000l\000i\000f\000i\000e\000d\000D\000i\000j\000k\000s\000t\000r\000a\000V\0002}{subsection.50}% 38
+\BOOKMARK [2][]{subsection.57}{\376\377\0009\000.\0004\000\040\000E\000v\000a\000l\000u\000a\000t\000i\000o\000n\000s}{section.38}% 39
+\BOOKMARK [1][]{section.58}{\376\377\0001\0000\000\040\000C\000o\000n\000c\000l\000u\000s\000i\000o\000n\000s\000\040\000a\000n\000d\000\040\000F\000u\000t\000u\000r\000e\000\040\000W\000o\000r\000k}{}% 40
+\BOOKMARK [2][]{subsection.59}{\376\377\0001\0000\000.\0001\000\040\000H\000o\000w\000\040\000t\000h\000e\000\040\000t\000e\000a\000m\000\040\000w\000o\000r\000k\000\040\000w\000e\000n\000t}{section.58}% 41
+\BOOKMARK [2][]{subsection.60}{\376\377\0001\0000\000.\0002\000\040\000W\000h\000a\000t\000\040\000w\000e\000\040\000h\000a\000v\000e\000\040\000l\000e\000a\000r\000n\000e\000d}{section.58}% 42
+\BOOKMARK [2][]{subsection.61}{\376\377\0001\0000\000.\0003\000\040\000I\000d\000e\000a\000s\000\040\000f\000o\000r\000\040\000t\000h\000e\000\040\000f\000u\000t\000u\000r\000e\000\040\000d\000e\000v\000e\000l\000o\000p\000m\000e\000n\000t\000\040\000o\000f\000\040\000y\000o\000u\000r\000\040\000a\000p\000p\000l\000i\000c\000a\000t\000i\000o\000n\000,\000\040\000n\000e\000w\000\040\000a\000l\000g\000o\000r\000i\000t\000h\000m\000s}{section.58}% 43
