+++ documentation/_minted-Snake-AI-Doc/5AFBEAB5275BA1621D1D7E7600D6EF4DAAFE4086999806E7A45FFDC085C11537.pygtex
@@ -1,0 +1,9 @@
+\begin{Verbatim}[commandchars=\\\{\}]
+\PYG{n}{Direction}\PYG{o}{[]}\PYG{+w}{ }\PYG{n}{notLosing}\PYG{+w}{ }\PYG{o}{=}\PYG{+w}{ }\PYG{n}{Arrays}\PYG{p}{.}\PYG{n+na}{stream}\PYG{p}{(}\PYG{n}{validMoves}\PYG{p}{)}
+\PYG{p}{.}\PYG{n+na}{filter}\PYG{p}{(}\PYG{n}{d}\PYG{+w}{ }\PYG{o}{\PYGZhy{}\PYGZgt{}}\PYG{+w}{ }\PYG{n}{head}\PYG{p}{.}\PYG{n+na}{moveTo}\PYG{p}{(}\PYG{n}{d}\PYG{p}{).}\PYG{n+na}{inBounds}\PYG{p}{(}\PYG{n}{mazeSize}\PYG{p}{))}
+\PYG{p}{.}\PYG{n+na}{filter}\PYG{p}{(}\PYG{n}{d}\PYG{+w}{ }\PYG{o}{\PYGZhy{}\PYGZgt{}}\PYG{+w}{ }\PYG{o}{!}\PYG{n}{opponent}\PYG{p}{.}\PYG{n+na}{elements}\PYG{p}{.}\PYG{n+na}{contains}\PYG{p}{(}\PYG{n}{head}\PYG{p}{.}\PYG{n+na}{moveTo}\PYG{p}{(}\PYG{n}{d}\PYG{p}{)))}
+\PYG{p}{.}\PYG{n+na}{filter}\PYG{p}{(}\PYG{n}{d}\PYG{+w}{ }\PYG{o}{\PYGZhy{}\PYGZgt{}}\PYG{+w}{ }\PYG{o}{!}\PYG{n}{snake}\PYG{p}{.}\PYG{n+na}{elements}\PYG{p}{.}\PYG{n+na}{contains}\PYG{p}{(}\PYG{n}{head}\PYG{p}{.}\PYG{n+na}{moveTo}\PYG{p}{(}\PYG{n}{d}\PYG{p}{)))}
+\PYG{p}{.}\PYG{n+na}{filter}\PYG{p}{(}\PYG{n}{d}\PYG{+w}{ }\PYG{o}{\PYGZhy{}\PYGZgt{}}\PYG{+w}{ }\PYG{o}{!}\PYG{n}{possiblePositionsOpponent}\PYG{p}{.}\PYG{n+na}{contains}\PYG{p}{(}\PYG{n}{head}\PYG{p}{.}\PYG{n+na}{moveTo}\PYG{p}{(}\PYG{n}{d}\PYG{p}{)))}
+\PYG{p}{.}\PYG{n+na}{sorted}\PYG{p}{()}
+\PYG{p}{.}\PYG{n+na}{toArray}\PYG{p}{(}\PYG{n}{Direction}\PYG{o}{[]}\PYG{p}{::}\PYG{k}{new}\PYG{p}{);}
+\end{Verbatim}
